+++ invenio_rdm_records/records/api.py
@@ -184,7 +184,7 @@ class CommonFieldsMixin:
         ),
         subjects=PIDListRelation(
             "metadata.subjects",
-            keys=["subject", "scheme", "props"],
+            keys=["subject", "scheme", "props", "identifiers"],
             pid_field=Subject.pid,
             cache_key="subjects",
         ),
@@ -536,9 +536,39 @@ class RDMRecord(CommonFieldsMixin, Record):
         published yet or all versions are deleted.
         """
         latest_record = cls.get_latest_by_parent(parent)
-        if latest_record.deletion_status != RecordDeletionStatusEnum.PUBLISHED.value:
+        if (
+            latest_record
+            and latest_record.deletion_status
+            != RecordDeletionStatusEnum.PUBLISHED.value
+        ):
             return None
         return latest_record
+
+    @classmethod
+    def get_previous_published_by_parent(cls, parent):
+        """Get the previous of latest published record for the specified parent record.
+
+        It might return None if there is no latest published version i.e not
+        published yet or all versions are deleted or there is only one published record.
+
+        This method is needed instead of `get_latest_published_by_parent` because during
+        publish the version state is updated before the record is actually published.
+        That means, that `get_latest_published_by_parent` returns always the record that
+        is about to be published and thus, we cannot use it in the `component.publish()`
+        method to retrieve the actual last published record.
+
+        Check `services.components.pids.PIDsComponent.publish()` for how it is used.
+        """
+        # We need no_autoflush because the record.versions access triggers automatically
+        # one
+        with db.session.no_autoflush:
+            records = cls.get_records_by_parent(parent)
+            for record in records:
+                latest_version_index = record.versions.latest_index
+                if latest_version_index > 1:
+                    if record.versions.index == latest_version_index - 1:
+                        return record
+            return None
 
 
 RDMFileRecord.record_cls = RDMRecord