+++ CHANGES.rst
@@ -12,6 +12,54 @@
 Changes
 =======
 
+Version v16.5.0 (released 2024-12-16)
+
+- pids: add support for optional DOI
+
+Version v16.4.1 (released 2024-12-11)
+
+- mappings: add missing `identifiers` to community orgs
+    * Adds the missing `identifiers` mapping field to community organizations
+
+Version v16.4.0 (released 2024-12-10)
+
+- bibtex: add trailing comma in url field
+- community-records: allow scan search
+    * Adds `scan` and `scan_params` arguments to
+      `CommunityRecordsService.search(...)`, to allow for serving scan
+      results (but only via the service).
+- serializer: updated subjects and affiliations in dcat
+- schema: added identifiers to subjects
+- serializers: add datapackage serializer (#1742)
+
+Version v16.3.4 (released 2024-12-06)
+
+- github: return None for `NOASSERTION` license
+- datacite: fix funding serialization for optional award fields
+    * Makes sure that we handle missing values for optional award fields
+      like "title" and "number".
+
+Version v16.3.3 (released 2024-12-04)
+
+- github: handle missing repo license
+
+Version v16.3.2 (released 2024-12-04)
+
+- github: lower license spdx id
+
+Version v16.3.1 (released 2024-12-02)
+
+- deposit-ui: make sure we handle null/undefined for SchemaField
+- deposit-ui: skip unecessary removal of empty values in serialization
+    * This initial removal of empty values can be dangerous, since the
+      `record` at this point is a UI object representation that could
+      potentially include circular references or very deeply nested objects.
+      Since `_removeEmptyValues` is recursive this can lead to stack
+      overflow errors.
+- deposit-ui: log errors on all deposit form actions
+    * This can help with debugging unexpected non-network related errors
+      that might occur in the logic before/after a REST API requests.
+
 Version v16.3.0 (released 2024-11-27)
 
 - github: added default license from Github API